+++ bob/database/Event.py
@@ -1,0 +1,7 @@
+class Event(Model):
+  iteration = ForeignKeyField(Iteration, related_name='events')
+  date = DateTimeField()
+  text = TextField()
+  tweet_id = CharField(null=True)
+  class Meta:
+    database = db
